rename and move worker, imports management, clean up logs
