+++ src/Components/ui/credits.tsx
@@ -2,7 +2,7 @@ import { Share } from "@/Components"
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "@/Components/ui"
 import { meta } from "@/assets/constants/meta"
 import { LogoIcon } from "@/assets/icons/logoIcon"
-import { useTranslation } from "@/common/utils/getLanguage"
+import { useTranslation } from "@/common/hooks/useLanguage"
 import "@/styles/Credits.css"
 import "@/styles/ui/Tooltip.css"
