+++ src/common/utils/getLanguage.tsx
@@ -3,30 +3,7 @@ import { en, fi } from "@/assets/texts"
 import { Language, type TranslationKey } from "@/models/translations"
 import { Fragment } from "react"
 
-import { FormContext, FormDispatchContext } from "@/common/providers/FormProvider"
-import { setLanguage } from "@/services/reducers/formReducer"
-import { useContext } from "react"
-
-export const useLanguage = () => {
-  const { language } = useContext(FormContext).formState
-  return { language, setLanguage: setNewLanguage }
-}
-
-function setNewLanguage(newLanguage: Language) {
-  const { dispatch } = useContext(FormDispatchContext)
-  return dispatch(setLanguage(newLanguage))
-}
-
-export const useTranslation = () => {
-  const { language } = useLanguage()
-
-  const t = (key: TranslationKey, placeholders?: Record<PropertyKey, string>) => {
-    return translate(language, key, placeholders)
-  }
-  return { t }
-}
-
-const translate = (
+export const translate = (
   language: Language,
   key: TranslationKey,
   placeholders?: Record<PropertyKey, string>,
