+++ src/services/initWorker.ts
@@ -1,0 +1,3 @@
+export const worker = new Worker(new URL("@/services/zxcvbn.worker.ts", import.meta.url), {
+  type: "module",
+})
