+++ src/styles/ui/RadioGroup.css
@@ -22,7 +22,7 @@
   width: fit-content;
   display: flex;
   align-items: center;
-  gap: 0.25rem;
+  gap: 0.5rem;
 }
 
 .RadioGroupItemLabel[aria-disabled="true"] {
